Small updates: code restructuring, standards, other lose ends.

+++ install/forum_data/src/login.php.t
@@ -299,6 +299,6 @@ function error_check()
 {TEMPLATE: LOGIN_PAGE}
 <?php
 	while (@ob_end_flush());
-	/* Clear old sessions. */
-	q('DELETE FROM {SQL_TABLE_PREFIX}ses WHERE time_sec<'.(__request_timestamp__- ($FUD_OPT_3 & 1 ? $SESSION_TIMEOUT : $COOKIE_TIMEOUT)));
+	/* Clear expired sessions AND anonymous sessions older than 1 day. */
+	q('DELETE FROM {SQL_TABLE_PREFIX}ses WHERE time_sec<'.(__request_timestamp__- ($FUD_OPT_3 & 1 ? $SESSION_TIMEOUT : $COOKIE_TIMEOUT)) .' OR (user_id>2000000000 AND time_sec<'.(__request_timestamp__- 86400).')');
 ?>
